--- a/olaf/_internals/resources/ecss.py
+++ b/olaf/_internals/resources/ecss.py
@@ -1,12 +1,13 @@
 """Resource for ECSS CANBus Extended Protocal standards"""
 
+import logging
 from os import geteuid
 from time import CLOCK_REALTIME, clock_settime, time
 
-from loguru import logger
-
 from ...common.ecss import scet_int_from_time, scet_int_to_time, utc_int_from_time, utc_int_to_time
 from ...common.resource import Resource
+
+logger = logging.getLogger(__file__)
 
 
 class EcssResource(Resource):