--- a/olaf/_internals/rest_api/templates/root.html
+++ b/olaf/_internals/rest_api/templates/root.html
@@ -4,7 +4,7 @@
   <style>
     #headerGrid {
       display: grid;
-      grid-template-columns: 25% 50% 25%;
+      grid-template-columns: 35% 30% 35%;
     }
     #headerCenter {
       text-align: center;