--- a/olaf/common/__init__.py
+++ b/olaf/common/__init__.py
@@ -20,6 +20,10 @@ def natsorted(data: list[str], ignore_case: bool = False) -> list[str]:
         List of natural sorted strings.
     """
 
-    convert = lambda text: int(text) if text.isdigit() else (text.lower() if ignore_case else text)
-    alphanum_key = lambda key: [convert(c) for c in re.split("([0-9]+)", key)]
+    def convert(text):
+        return int(text) if text.isdigit() else (text.lower() if ignore_case else text)
+
+    def alphanum_key(key):
+        return [convert(c) for c in re.split("([0-9]+)", key)]
+
     return sorted(data, key=alphanum_key)