--- a/olaf/common/oresat_file_cache.py
+++ b/olaf/common/oresat_file_cache.py
@@ -7,7 +7,6 @@ from os.path import abspath, basename, isfile
 from pathlib import Path
 from threading import Lock
 
-from . import natsorted
 from .oresat_file import OreSatFile
 
 
@@ -37,7 +36,7 @@ class OreSatFileCache:
                 self._data.append(oresat_file)
             except Exception:  # pylint: disable=W0718
                 remove(self._dir + f)  # invalid file name
-        self._data = natsorted(self._data)
+        self._data = sorted(self._data)
 
     def __len__(self) -> int:
         with self._lock:
@@ -78,7 +77,7 @@ class OreSatFileCache:
 
             if not overwrite:
                 self._data.append(oresat_file)
-                self._data = natsorted(self._data)
+                self._data = sorted(self._data)
 
     def remove(self, file_name: str):
         """Remove a file from cache