--- a/pyproject.toml
+++ b/pyproject.toml
@@ -21,7 +21,6 @@ classifiers = [
 dependencies = [
     "canopen",
     "flask",
-    "loguru",
     "natsort",
     "psutil",
     "oresat-configs",
@@ -63,13 +62,16 @@ linters = "pycodestyle,pyflakes,pylint,mccabe,mypy,radon"
 # W0102:    Dangerous default value as argument
 # W0107:    Unnecessary pass statement
 # R0903:    Too few public methods
-ignore = "E402,C901,C0103,E203,R0912,R0915,R901,R901,R0914,C0413,C0206,R1716,W1514,R0902,R0913,W0707,W0102,W0107,R0903"
+# W1201:    Use % formatting in logging functions
+# W1203:    Use %s formatting in logging functions
+ignore = "E402,C901,C0103,E203,R0912,R0915,R901,R901,R0914,C0413,C0206,R1716,W1514,R0902,R0913,W0707,W0102,W0107,R0903,W1201,W1203"
 max_line_length = 100
 
 [[tool.pylama.files]]
 path = "*/__init__.py"
 # W0611:    Imported but unused
-ignore = "W0611"
+# W1201:    Use % formatting in logging functions
+ignore = "W0611,W1201"
 
 [[tool.pylama.files]]
 path = "tests/*"