--- a/requirements.txt
+++ b/requirements.txt
@@ -3,7 +3,6 @@ build
 canopen
 flask
 isort
-loguru
 natsort
 oresat-configs
 psutil