--- a/tests/__init__.py
+++ b/tests/__init__.py
@@ -1,0 +1,4 @@
+import logging
+
+# this will stop logging to console
+logging.disable(logging.CRITICAL)