--- a/tests/internals/services/__init__.py
+++ b/tests/internals/services/__init__.py
@@ -4,10 +4,8 @@ import can
 import canopen
 from oresat_configs import OreSatConfig, OreSatId
 
-from olaf import OreSatFileCache, Service, logger
+from olaf import OreSatFileCache, Service
 from olaf._internals.node import Node
-
-logger.disable("olaf")
 
 
 class MockNode(Node):