Tested with `python -X importtime run.py gps -b can0` on C3 card at max CPU frequency

Import time before changes: 5.9s
Import time after changes: 4.1s

Changes:
- Replaced `natsort` with a custom function (removed 300ms)
- Replaced `loguru` with build-in `logging` module (removed 260ms)
- Set minimum `python-can` to 4.3.0 (release fixed internal `pkg_resource` slowdown, removed 1s)

Could remove 1.5s more if removed or split out the`Flask` app, but that could make the projects harder to modify for new students.